Split mock code into own file
